5.12

+++ src/main/java/com/zllUserCenter/findfriendbackend/model/domain/UserTeam.java
@@ -27,7 +27,7 @@ public class UserTeam {
     /**
      * 队伍id
      */
-    private Integer teamId;
+    private Long teamId;
 
     /**
      * 加入时间
